Remove ES 1.x search backend
